packages/chatbot-server/pyproject.toml
[project]
name = "chatbot-server"
version = "0.1.0"
description = ""
authors = [
    {name = "JesuFrancesco",email = "117495071+JesuFrancesco@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "oracledb (>=3.1.1,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "chatbot_server", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
